Pluggable Replicator
New Workflow for replication

                                Client

                                  ||
                                  ||
                                  ||
                                  \/

                                Request

                                  ||
                                  ||
                                  ||
                                  \/

                            HandleOrForward

                               /\    ||
[Header:REPLICATION_REQUEST]   ||    ||  Header: ELIGIBLE_NODES=[]
 (times: nrOfReplica)          ||    ||
                               ||    \/

                            ReplicationRouter

HandleOrForward Interceptor always sets a header called "ELIGIBLE_NODES" which contains list of all nodes that're eligible to participate in the current operation as defined by RoutingStrategy.
SequentialReplicator would know to replicate the incomingMessage based on ELIGIBLE_NODES, if REPLICATION_REQUEST header is not set. Now while replicating the message it would add REPLICATION_REQUEST header to those messages so ReplicationRouter on the other node (or process) will not forward the message instead just delegate it to the local handler.

As part of this change we've also made ReplicationRouter an abstract class and provided an implementation called SequentialReplicator which does Sequential replication to all the ELIGIBLE_NODES.