Switching to a different codecoverage plugin
There was an issue with the cobertura maven plugin wherein it was re-running the tests twice, which means with this change we could reduce the build time by almost half assuming the artifacts are cached on SnapCI.

With the new plugin it reduced the coverage by half because it was taking into account the gRPC generated code as well. Hence had to move the generated code to a new package and exclude it specifically in the plugin configuration. I wish life was much easier.

Ref - http://stackoverflow.com/a/31387173/1058334