MemberListener + Membership contract changes
TLDR;
- Change contracts in Membership abstract class with scaladocs
- Adding MemberListener
- Bootstrapper -> SeedProvider - changed the names as it was confusing to use
- Integrated the new Membership contract with Atomix + ScaleCube implementations + fixing the tests too

MemberListener are pluggable event listeners to changes in membership. This allows us to write composable components like how should ConsistentHashRing react to new nodes or exit of an existing nodes in the cluster.

Membership implementations should just plug `onJoin` and `onLeave` methods in Membership to the actual membership events in the cluster. Membership would take care of calling all the event listeners in order with MemberAddress of the node that has the respective event.